requirements.txt
Django==3.0.8
djangorestframework==3.11.1
python-telegram-bot==12.8
psycopg2==2.8.5
pygments==2.6.1